added limit to radius, seperate drawer helper files and reverted default font
