+++ .gitignore
@@ -1,2 +1,3 @@
 /osk-sdl
 /go.sh
+.vscode
