+++ .vscode/settings.json
@@ -1,5 +1,8 @@
 {
     "files.associations": {
-        "*.tcc": "cpp"
+        "*.tcc": "cpp",
+        "string_view": "cpp",
+        "cerrno": "cpp",
+        "system_error": "cpp"
     }
 }
