+++ osk.conf
@@ -5,7 +5,7 @@ wallpaper = #0099FF
 
 keyboard-background = #333333;
 keyboard-map = us
-keyboard-font = /usr/share/fonts/TTF/DejaVuSans.ttf
+keyboard-font = /usr/share/fonts/ttf-dejavu/DejaVuSans.ttf
 
-key-radius = 20
-inputbox-radius = 20
+key-radius = 10
+inputbox-radius = 35
